Updated license. Please accept and merge.
OK

+++ app/src/main/java/com/njlabs/showjava/processor/JarExtractor.java
@@ -3,16 +3,24 @@ package com.njlabs.showjava.processor;
 import android.util.Log;
 
 import com.crashlytics.android.Crashlytics;
+import com.googlecode.dex2jar.Method;
+import com.googlecode.dex2jar.ir.IrMethod;
 import com.googlecode.dex2jar.reader.DexFileReader;
 import com.googlecode.dex2jar.v3.Dex2jar;
-import com.njlabs.showjava.Constants;
+import com.googlecode.dex2jar.v3.DexExceptionHandler;
+import com.njlabs.showjava.utils.StringUtils;
+import com.njlabs.showjava.utils.logging.Ln;
 
 import org.jf.dexlib2.DexFileFactory;
 import org.jf.dexlib2.iface.ClassDef;
 import org.jf.dexlib2.iface.DexFile;
 import org.jf.dexlib2.immutable.ImmutableDexFile;
+import org.objectweb.asm.tree.MethodNode;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.List;
@@ -23,7 +31,7 @@ import java.util.List;
 @SuppressWarnings({"ResultOfMethodCallIgnored", "ConstantConditions"})
 public class JarExtractor extends ProcessServiceHelper {
 
-    ArrayList<String> ignoredLibs;
+    private final ArrayList<String> ignoredLibs;
 
     public JarExtractor(ProcessService processService) {
         this.processService = processService;
@@ -34,6 +42,12 @@ public class JarExtractor extends ProcessServiceHelper {
         this.sourceOutputDir = processService.sourceOutputDir;
         this.javaSourceOutputDir = processService.javaSourceOutputDir;
         ignoredLibs = new ArrayList<>();
+
+        //////
+        printStream = new PrintStream(new ProgressStream());
+        System.setErr(printStream);
+        System.setOut(printStream);
+        //////
     }
 
     public void extract() {
@@ -42,18 +56,21 @@ public class JarExtractor extends ProcessServiceHelper {
         Runnable runProcess = new Runnable() {
             @Override
             public void run() {
+                loadIgnoredLibs();
                 apkToDex();
-                dexToJar();
+                if(!processService.decompilerToUse.equals("jadx")){
+                    dexToJar();
+                }
                 startJavaExtractor();
             }
         };
-        Thread extractionThread = new Thread(group, runProcess, "DEX TO JAR EXTRACTION", Constants.STACK_SIZE);
+        Thread extractionThread = new Thread(group, runProcess, "DEX TO JAR EXTRACTION", processService.STACK_SIZE);
         extractionThread.setPriority(Thread.MAX_PRIORITY);
         extractionThread.setUncaughtExceptionHandler(exceptionHandler);
         extractionThread.start();
     }
 
-    public void apkToDex() {
+    private void apkToDex() {
         DexFile dexFile = null;
         try {
             dexFile = DexFileFactory.loadDexFile(packageFilePath, 19);
@@ -65,49 +82,7 @@ public class JarExtractor extends ProcessServiceHelper {
         broadcastStatus("optimising", "");
 
         for (ClassDef classDef : dexFile.getClasses()) {
-            if (
-                    !classDef.getType().startsWith("Lcom/google/apps/")
-                            && !classDef.getType().startsWith("Landroid/")
-                            && !classDef.getType().startsWith("Lcom/android/")
-                            && !classDef.getType().startsWith("Lcom/google/android/gms/")
-                            && !classDef.getType().startsWith("Lcom/google/common/")
-                            && !classDef.getType().startsWith("Lcom/google/auto/")
-                            && !classDef.getType().startsWith("Lcom/google/ads/")
-                            && !classDef.getType().startsWith("Lcom/google/android/vending/")
-
-                            && !classDef.getType().startsWith("Lcom/squareup/okhttp")
-                            && !classDef.getType().startsWith("Lcom/google/gson")
-                            && !classDef.getType().startsWith("Lcom/square/picasso")
-                            && !classDef.getType().startsWith("Lcom/nineoldandroids")
-                            && !classDef.getType().startsWith("Lbolts")
-                            && !classDef.getType().startsWith("Lcom/mikepenz/iconics")
-                            && !classDef.getType().startsWith("Lretrofit")
-                            && !classDef.getType().startsWith("Lorg/parceler/")
-                            && !classDef.getType().startsWith("Lbutterknife")
-                            && !classDef.getType().startsWith("Lcom/loopj/android/")
-                            && !classDef.getType().startsWith("Lorg/objectweb/asm/")
-                            && !classDef.getType().startsWith("Lcom/crashlytics/")
-                            && !classDef.getType().startsWith("Lio/fabric/sdk/")
-                            && !classDef.getType().startsWith("Lcom/androidquery/")
-                            && !classDef.getType().startsWith("Lcom/parse/")
-                            && !classDef.getType().startsWith("Lcom/actionbarsherlock/")
-                            && !classDef.getType().startsWith("Lorg/apache/")
-                            && !classDef.getType().startsWith("Lorg/acra/")
-                            && !classDef.getType().startsWith("Ljavax/")
-                            && !classDef.getType().startsWith("Lorg/joda/")
-                            && !classDef.getType().startsWith("Lorg/antlr/")
-                            && !classDef.getType().startsWith("Ljunit/")
-                            && !classDef.getType().startsWith("Lorg/codehaus/jackson/")
-                            && !classDef.getType().startsWith("Lcom/fasterxml/")
-                            && !classDef.getType().startsWith("Lnet/sourceforge/")
-                            && !classDef.getType().startsWith("Lorg/achartengine/")
-                            && !classDef.getType().startsWith("Lcom/bugsense/")
-                            && !classDef.getType().startsWith("Lorg/andengine/")
-                            && !classDef.getType().startsWith("Lcom/inmobi/")
-                    )
-
-            {
-
+            if (!isIgnored(classDef.getType())) {
                 final String CurrentClass = classDef.getType();
                 broadcastStatus("optimising_class", CurrentClass.replaceAll("Processing ", ""));
                 classes.add(classDef);
@@ -132,9 +107,11 @@ public class JarExtractor extends ProcessServiceHelper {
             broadcastStatus("exit");
             UIHandler.post(new ToastRunnable("The app you selected cannot be decompiled. Please select another app."));
         }
+
+
     }
 
-    public void dexToJar() {
+    private void dexToJar() {
         Log.i("STATUS", "Jar Extraction Started");
 
         broadcastStatus("dex2jar");
@@ -148,32 +125,80 @@ public class JarExtractor extends ProcessServiceHelper {
         boolean printIR = false; // print ir to System.out
         boolean optimizeSynchronized = true; // Optimise-synchronised
 
-        //////
-        PrintStream printStream = new PrintStream(new ProgressStream());
-        System.setErr(printStream);
-        System.setOut(printStream);
-        //////
-
         File PerAppWorkingDirectory = new File(sourceOutputDir);
         File file = new File(PerAppWorkingDirectory + "/" + packageName + ".jar");
 
-        try {
-            DexFileReader reader = new DexFileReader(new File(PerAppWorkingDirectory + "/optimised_classes.dex"));
-            Dex2jar.from(reader).reUseReg(reuseReg).topoLogicalSort(topologicalSort || topologicalSort1).skipDebug(!debugInfo)
-                    .optimizeSynchronized(optimizeSynchronized).printIR(printIR).verbose(verbose).to(file);
-        } catch (Exception e) {
-            Crashlytics.logException(e);
-            broadcastStatus("exit_process_on_error");
+        File dexFile = new File(PerAppWorkingDirectory + "/optimised_classes.dex");
+
+        if (dexFile.exists() && dexFile.isFile()) {
+            DexExceptionHandlerMod dexExceptionHandlerMod = new DexExceptionHandlerMod();
+            try {
+                DexFileReader reader = new DexFileReader(dexFile);
+                Dex2jar dex2jar = Dex2jar.from(reader).reUseReg(reuseReg).topoLogicalSort(topologicalSort || topologicalSort1).skipDebug(!debugInfo)
+                        .optimizeSynchronized(optimizeSynchronized).printIR(printIR).verbose(verbose);
+                dex2jar.setExceptionHandler(dexExceptionHandlerMod);
+                dex2jar.to(file);
+            } catch (Exception e) {
+                broadcastStatus("exit_process_on_error");
+            }
+
+            Log.i("STATUS", "Clearing cache");
+            File ClassDex = new File(PerAppWorkingDirectory + "/optimised_classes.dex");
+            ClassDex.delete();
         }
 
-        Log.i("STATUS", "Clearing cache");
-        File ClassDex = new File(PerAppWorkingDirectory + "/optimised_classes.dex");
-        ClassDex.delete();
     }
 
     private void startJavaExtractor() {
         JavaExtractor javaExtractor = new JavaExtractor(processService);
         javaExtractor.extract();
+    }
+
+    private void loadIgnoredLibs() {
+        String ignoredList = (processService.IGNORE_LIBS ? "ignored.list":"ignored_basic.list");
+        BufferedReader reader = null;
+        try {
+            reader = new BufferedReader(new InputStreamReader(processService.getAssets().open(ignoredList)));
+            String mLine = reader.readLine().trim();
+            while (mLine != null) {
+                mLine = mLine.trim();
+                if (mLine.length() != 0) {
+                    ignoredLibs.add(StringUtils.toClassName(mLine));
+                }
+                mLine = reader.readLine();
+            }
+        } catch (IOException e) {
+            Crashlytics.logException(e);
+        } finally {
+            if (reader != null) {
+                try {
+                    reader.close();
+                } catch (IOException e) {
+                    Crashlytics.logException(e);
+                }
+            }
+        }
+    }
+
+    private boolean isIgnored(String className) {
+        for (String ignoredClass : ignoredLibs) {
+            if (className.startsWith(ignoredClass)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private class DexExceptionHandlerMod implements DexExceptionHandler {
+        @Override
+        public void handleFileException(Exception e) {
+            Ln.d("Dex2Jar Exception " + e);
+        }
+
+        @Override
+        public void handleMethodTranslateException(Method method, IrMethod irMethod, MethodNode methodNode, Exception e) {
+            Ln.d("Dex2Jar Exception " + e);
+        }
     }
